Revert "Merge pull request #1214 from Crown-Commercial-Service/bugfix/CAS-0000_Issue_with_enity_name"
This reverts commit 2bfe0e5b0e856742bc8148ca0e13827b6e286a99, reversing
changes made to dcc038f9c627670962db731a41e94f9848e324fa.

+++ src/main/java/uk/gov/crowncommercial/dts/scale/cat/model/entity/ca/TaskEntity.java
@@ -30,7 +30,7 @@ public class TaskEntity {
   private String name;
 
   @Column(name="group_id")
-  private Integer groupId;
+  private Long groupId;
 
   @Column(name = "task_object")
   private String recordType;
